Erase fluid_pre_refining_mesher and use fluid_mesher
